+++ mobile-client-android/sentinel_android/app/src/main/java/sentinelgroup/io/sentinel/network/model/Usage.java
@@ -3,7 +3,7 @@ package sentinelgroup.io.sentinel.network.model;
 import java.util.List;
 
 public class Usage {
-    public int due;
+    public long due;
     public Stats stats;
     public List<Session> sessions = null;
 }
